Actualización de correción de errores

+++ src/apps/proyectos/view.py
@@ -5,7 +5,6 @@ import base64
 from flask import Flask, json, render_template, request, redirect, url_for, send_from_directory, flash, jsonify
 from flask_mysqldb import MySQL
 from flask_wtf.csrf import CSRFProtect
-from flask_socketio import SocketIO
 from flask_login import LoginManager, current_user
 
 
@@ -19,8 +18,6 @@ app = Flask(__name__)
 lenguajes=Lenguajes()
 
 csrf = CSRFProtect() # Protección de csrf pa los form's
-
-socketio = SocketIO(app) # Envío en tiempo real para comentarios
 
 db = MySQL(app)
 
@@ -44,22 +41,14 @@ login_manager.login_view = 'login'
 # --------------------------------------------------------------------
 # --------------------------------------------------------------------
 
-# def get_base_query():
-#     return (
-#     'SELECT proyectos.*, usuarios.Foto, usuarios.NombreUsuario AS NombreUsuario, '
-#     'IFNULL(AVG(calificacion.Calificacion), 0) AS PromedioCalificacion '
-#     'FROM proyectos '
-#     'JOIN usuarios ON proyectos.IDUsuario = usuarios.ID '
-#     'LEFT JOIN calificacion ON proyectos.ID = calificacion.IDProyecto '
-#     )
 
 def get_base_query2():
     return (
-    'SELECT p.ID, p.IDUsuario, p.NombreProyecto, p.DescripcionProyecto, p.Fecha, p.Imagen, usuarios.Foto, usuarios.NombreUsuario AS NombreUsuario, '
-    'IFNULL(AVG(calificacion.Calificacion), 0) AS PromedioCalificacion '
-    'FROM PROYECTOS P '
-    'JOIN usuarios ON p.IDUsuario = usuarios.ID '
-    'LEFT JOIN calificacion ON p.ID = calificacion.IDProyecto '
+    'SELECT p.ID, p.IDUsuario, p.NombreProyecto, p.DescripcionProyecto, p.Fecha, p.Imagen, Usuarios.Foto, Usuarios.NombreUsuario AS NombreUsuario, '
+    'IFNULL(AVG(Calificacion.Calificacion), 0) AS PromedioCalificacion '
+    'FROM Proyectos p '
+    'JOIN Usuarios ON p.IDUsuario = Usuarios.ID '
+    'LEFT JOIN Calificacion ON p.ID = Calificacion.IDProyecto '
     )
 
 def order_base(order_by, base):
@@ -70,7 +59,7 @@ def order_base(order_by, base):
         base += 'ORDER BY PromedioCalificacion DESC'
     else:
         base += 'ORDER BY RAND()'  # Ordenamiento aleatorio
-    
+
     return base
 
 
@@ -99,10 +88,10 @@ def structure_data_Projects(objetoProyecto):
         # Convierte la imagen en un formato correcto para enviar una cadena base64 válida
         if 'Imagen' in proyecto:
             proyecto['Imagen'] = convert_image_base64(proyecto['Imagen'])
-            
+
         if isinstance(proyecto['Archivo'], bytes):
                 proyecto['Archivo'] = proyecto['Archivo'].decode('utf-8') # Decodificar los bytes de los archivos a una cadena
-        
+
         if 'Archivo' in proyecto:
             # Pregunta si existen archivos y los separa
             if proyecto['Archivo'] is not None:
@@ -138,7 +127,7 @@ def convert_image_base64(byte_image):
 def get_image_comment(objetoComentarios):
     for comentario in objetoComentarios:
         comentario['Foto'] = get_photo_perfil(comentario['Foto'])
-    
+
     return objetoComentarios
 
 
@@ -148,9 +137,9 @@ def extract_info_query(query, param=None):
         cursor.execute(query, param)
     else:
         cursor.execute(query)
-    
+
     resultado = cursor.fetchall()
-    
+
     nameColumn = [column[0] for column in cursor.description]
 
     cursor.close()
@@ -173,9 +162,9 @@ def extract_dataUser_comments(query, param):
 
 def get_photo_perfil(str_foto):
     if str_foto:
-        return f'static/images/perfil/{str_foto}'
+        return f'static/img/perfil/{str_foto}'
     else:
-        return f'static/images/perfil/default.jpg'
+        return f'static/img/perfil/default.jpg'
 
 
 
@@ -197,9 +186,9 @@ def save_comment(id_project, comentario):
     formatedDate = fecha.strftime('%Y-%m-%d %H:%M:%S')
 
     cursor = db.connection.cursor()
-    sql = "INSERT INTO comentarios (IDProyecto,IDUsuario, Comentario, Fecha) VALUES (%s, %s, %s, %s)"
+    sql = "INSERT INTO Comentarios (IDProyecto,IDUsuario, Comentario, Fecha) VALUES (%s, %s, %s, %s)"
     data = (id_project, current_user.id, comentario, formatedDate)
-    
+
     cursor.execute(sql, data)
     db.connection.commit()
 
@@ -240,7 +229,7 @@ def calificar(id):
         # Resto del código
         if calificar_proyecto:
             cursor = db.connection.cursor()
-            sql = "INSERT INTO calificacion (IDProyecto,IDUsuario, Calificacion) VALUES (%s, %s, %s)"
+            sql = "INSERT INTO Calificacion (IDProyecto,IDUsuario, Calificacion) VALUES (%s, %s, %s)"
             data = (id, current_user.id, calificar_proyecto)
             cursor.execute(sql, data)
             db.connection.commit()
@@ -265,7 +254,7 @@ def re_calificar(id_pro,id_ca):
         # Resto del código
         if calificar_proyecto:
             cursor = db.connection.cursor()
-            sql = "UPDATE calificacion SET IDProyecto = %s, IDUsuario= %s, Calificacion = %s WHERE ID = %s"
+            sql = "UPDATE Calificacion SET IDProyecto = %s, IDUsuario= %s, Calificacion = %s WHERE ID = %s"
             data = (id_pro, current_user.id, calificar_proyecto,id_ca)
             cursor.execute(sql, data)
             db.connection.commit()
@@ -287,7 +276,7 @@ def idioma():
 
     if idioma:
         cursor = db.connection.cursor()
-        sql = "UPDATE usuarios SET IDLenguaje = %s WHERE ID = %s"
+        sql = "UPDATE Usuarios SET IDLenguaje = %s WHERE ID = %s"
         data = (idioma_select,current_user.id)
         cursor.execute(sql, data)
         db.connection.commit()
@@ -301,13 +290,13 @@ def idioma():
 #* ********************************************************************************************* *#
 
 def misProyectos(order_by='random', sessionTime=0):
-    
+
     # Query del projects
     base_query = get_base_query2() + ('WHERE p.IDUsuario = %s '
-                                    'GROUP BY p.ID, usuarios.NombreUsuario ')
-    
+                                    'GROUP BY p.ID, Usuarios.NombreUsuario ')
+
     base_query = order_base(order_by, base_query)
-    
+
     # Proceso para obtener la lista de los proyectos
     objetoProyecto = structure_data_ProjectPersonal(extract_info_query(base_query, (current_user.id, )))
 
@@ -331,7 +320,7 @@ def misProyectos(order_by='random', sessionTime=0):
 def allProyectos(order_by='random', sessionTime=0):
 
     #Para poder usar el filtro cambio 15-06-24
-    base_query = get_base_query2() + 'GROUP BY P.ID, usuarios.NombreUsuario '
+    base_query = get_base_query2() + 'GROUP BY p.ID, Usuarios.NombreUsuario '
 
     base_query = order_base(order_by, base_query)
 
@@ -340,7 +329,7 @@ def allProyectos(order_by='random', sessionTime=0):
 
     # # Proceso para definir el idioma del sistema
     idioma = getLanguajeSystem()
-    
+
     return render_template('proyectos/prueba.html',
                         idioma=idioma,
                         proyectos=objetoProyecto,
@@ -361,27 +350,27 @@ def consultProject(id_project):
 
         query_project = (
             'SELECT DescripcionProyecto, Fecha, Archivo '
-            'FROM PROYECTOS '
-            'WHERE PROYECTOS.ID = %s '
+            'FROM Proyectos '
+            'WHERE Proyectos.ID = %s '
         )
 
         query_comments = (
             'SELECT IDUsuario, Comentario, Fecha '
-            'FROM comentarios C '
+            'FROM Comentarios C '
             'WHERE IDProyecto = %s '
             'ORDER BY Fecha desc '
             )
-        
+
         query_comments_user = (
             'SELECT DISTINCT C.IDUsuario, U.NombreUsuario AS NombreUsuario, U.Foto '
-            'FROM comentarios C, usuarios U '
+            'FROM Comentarios C, Usuarios U '
             'WHERE C.IDUsuario = U.ID '
             'AND IDProyecto = %s '
         )
 
         query_rating_user =  (
             'SELECT ID, Calificacion '
-            'FROM CALIFICACION C '
+            'FROM Calificacion C '
             'WHERE IDProyecto = %s '
             'AND C.IDUsuario = %s '
         )
@@ -389,15 +378,18 @@ def consultProject(id_project):
         # Proceso para obtener el proyecto
         objetoProyecto = structure_data_Projects(extract_info_query(query_project, (id_project, )))
 
+        if not objetoProyecto[0]['Archivos'][0]:
+            objetoProyecto[0]['Archivos'] = 0
+
         # Proceso para obtener la lista de los comentarios de un proyecto
         objetoComentarios = extract_info_query(query_comments, (id_project, ))
-        
+
         # Proceso para obtener info de los usuarios que comentaron un proyecto
         users = extract_dataUser_comments(query_comments_user, (id_project, ))
 
         # Proceso para obtener la lista de los calificacion
         rating = extract_info_query(query_rating_user, (id_project, current_user.id,))
-        
+
         return jsonify({'message': 'Respuesta recibida',
                         'proyecto': objetoProyecto,
                         'comentarios': objetoComentarios,
@@ -425,7 +417,7 @@ def consultProject(id_project):
 #? ************************* VER EL ARCHIVO ************************* ?#
 #? ****************************************************************** ?#
 
-# Decorador de Flask que define una nueva ruta en la aplicación web. Cuando un usuario accede a una 
+# Decorador de Flask que define una nueva ruta en la aplicación web. Cuando un usuario accede a una
 # URL que coincide con este patrón, se ejecuta la función `mostrar_archivo`.
 
 def mostrar_archivo(nombre_usuario, nombre_proyecto, nombre_archivo):
@@ -446,12 +438,12 @@ def crear_proyecto():
     cursor.execute("SELECT ID, NombreUsuario FROM Usuarios") # Obtener los usuarios y sus IDs
     usuarios = cursor.fetchall() # Obtener todos los resultados de la consulta
     cursor.close()
-    
+
     # Obtener la fecha actual
     fecha_actual = datetime.now().strftime('%Y-%m-%d')
-    
+
     idioma = getLanguajeSystem()
-        
+
     return render_template('proyectos/crear_proyecto.html',idioma=idioma, usuarios=usuarios, fecha_actual=fecha_actual, usuario_actual=current_user.username)
 
 # Guardar los registros/proyectos
@@ -523,18 +515,18 @@ def delete(ID):
 
         for comentario_id in comentarios_ids:
             cursor.execute("DELETE FROM Comentarios WHERE ID = %s", (comentario_id,))
-        
+
         # Verificar si hay calificaciones relacionadas y eliminarlas
         cursor.execute("SELECT ID FROM Calificacion WHERE IDProyecto = %s", (ID,))
         calificaciones_ids = cursor.fetchall()
 
         for calificacion_id in calificaciones_ids:
             cursor.execute("DELETE FROM Calificacion WHERE ID = %s", (calificacion_id,))
-        
+
         # Obtener la información del proyecto antes de eliminarlo
-        cursor.execute("SELECT u.NombreUsuario, p.NombreProyecto, p.Archivo FROM proyectos p JOIN Usuarios u ON p.IDUsuario = u.ID WHERE p.ID = %s", (ID,))
+        cursor.execute("SELECT u.NombreUsuario, p.NombreProyecto, p.Archivo FROM Proyectos p JOIN Usuarios u ON p.IDUsuario = u.ID WHERE p.ID = %s", (ID,))
         result = cursor.fetchone()
-        
+
         if result:
             nombre_usuario, nombre_proyecto, archivos_combinados = result
 
@@ -546,7 +538,7 @@ def delete(ID):
                 shutil.rmtree(carpeta_proyecto)
 
             # Eliminar el registro del proyecto en la base de datos
-            cursor.execute("DELETE FROM proyectos WHERE ID = %s", (ID,))
+            cursor.execute("DELETE FROM Proyectos WHERE ID = %s", (ID,))
             db.connection.commit()
 
             # flash(f"El proyecto '{nombre_proyecto}' y sus comentarios/calificaciones han sido eliminados correctamente.", "success")
@@ -562,23 +554,23 @@ def delete(ID):
 # Ruta para editar un proyecto específico mediante su ID
 def editar_proyecto(ID):
     cursor = db.connection.cursor()
-    cursor.execute('SELECT * FROM proyectos WHERE ID=%s', (ID,))
+    cursor.execute('SELECT * FROM Proyectos WHERE ID=%s', (ID,))
     proyecto = cursor.fetchone()
-    
+
     # Obtener el nombre del usuario
     cursor.execute('SELECT NombreUsuario FROM Usuarios WHERE ID=%s', (proyecto[1],))
     nombre_usuario = cursor.fetchone()[0]
-    
+
     # Obtener la lista de archivos asociados al proyecto
     archivos_proyecto = proyecto[5].split(',') if proyecto[5] else []
-    
+
     cursor.close()
 
     # Convertir la imagen del proyecto a base64 para mostrarla en la interfaz de usuario
     if proyecto[6] is not None:
         imagen_base64 = base64.b64encode(proyecto[6]).decode('utf-8')
     else:
-        imagen_base64 = url_for('static', filename='images/perfil/default.jpg')
+        imagen_base64 = url_for('static', filename='img/perfil/default.jpg')
     # Convertir los detalles del proyecto en un diccionario para pasarlos a la plantilla
     proyecto_dict = {
         'ID': proyecto[0],
@@ -592,7 +584,7 @@ def editar_proyecto(ID):
     }
     # Proceso para definir el idioma del sistema
     idioma = getLanguajeSystem()
-    
+
     return render_template('proyectos/editar_proyecto.html', idioma=idioma, proyecto=proyecto_dict)
 
 # Ruta para actualizar un proyecto después de la edición
@@ -624,13 +616,19 @@ def actualizar_proyecto():
         carpeta_proyecto_nuevo = os.path.join(carpeta_usuario, nombre_proyecto_nuevo)
 
         # Eliminar archivos marcados para eliminar de la carpeta del proyecto
-        for index in removed_files:
-            archivo_a_eliminar = archivos_existentes[index]
-            ruta_archivo_a_eliminar = os.path.join(carpeta_proyecto_anterior, archivo_a_eliminar)
-            if os.path.exists(ruta_archivo_a_eliminar):
-                os.remove(ruta_archivo_a_eliminar)
-            # Eliminar el archivo de la lista de archivos existentes
-            del archivos_existentes[index]
+        if removed_files:
+            # Ordenar índices en orden descendente para eliminar correctamente
+            removed_files.sort(reverse=True)
+            for index in removed_files:
+                if index < len(archivos_existentes):
+                    print(index)
+                    print(archivos_existentes)
+                    archivo_a_eliminar = archivos_existentes[index]
+                    ruta_archivo_a_eliminar = os.path.join(carpeta_proyecto_anterior, archivo_a_eliminar)
+                    if os.path.exists(ruta_archivo_a_eliminar):
+                        os.remove(ruta_archivo_a_eliminar)
+                    # Eliminar el archivo de la lista de archivos existentes
+                    del archivos_existentes[index]
 
         # Renombrar la carpeta del proyecto si el nombre del proyecto ha cambiado
         if nombre_proyecto_nuevo != nombre_proyecto_anterior:
